refactor: Tabs 컴포넌트 export 방식 변경
